editor.js cleanup
