+++ app/views/editor.ejs
@@ -65,7 +65,7 @@
                                 <a class="dropdown-item" href="/projects">My Projects</a>
                                 <div class="dropdown-divider"></div>
                                 <a class="dropdown-item" href="./settings/share">Share</a>
-                                <span class="dropdown-item" style="cursor: pointer;" onclick="editor.compileLaTeX(true);">Compile</span>
+                                <span class="dropdown-item" style="cursor: pointer;" onclick="editor.CompileLaTeX(true);">Compile</span>
                                 <div class="dropdown-divider"></div>
                                 <a class="dropdown-item" href="./settings">Settings</a>
                             </div>
