+++ frontend/js/editor/AceCursorManager.js
