+++ frontend/js/editor/BootstrapManager.js
@@ -1,0 +1,12 @@
+const bootstrap = require("bootstrap");
+
+// Enable Tooltips
+$(() => {
+    $('[data-toggle="tooltip"]').tooltip()
+});
+
+// Menu Toggle Script
+$("#menu-toggle").on('click', (e) => {
+    e.preventDefault();
+    $("#wrapper").toggleClass("toggled");
+});
