+++ frontend/js/editor/index.js
@@ -1,0 +1,10 @@
+const AceManager = require("./AceManager");
+const TitleRename = require("./TitleRename");
+const SocketManager = require("./SocketManager");
+const LayoutManager = require("./LayoutManager");
+const CompileLaTeX = require("./CompileLaTeX");
+const BootstrapManager = require("./BootstrapManager");
+
+module.exports = {
+    CompileLaTeX: CompileLaTeX
+}
